ms-recommendation/requirements/development.txt
asgiref==3.11.1
Django==5.2
djangorestframework==3.16.1
psycopg2-binary==2.9.10
sqlparse==0.5.5
